ported xtend inferrer code to java
