+++ org.eclipse.xtext.xbase.testlanguages/build.properties
@@ -1,6 +1,5 @@
 source.. = src/,\
-          src-gen/,\
-          xtend-gen/
+          src-gen/
 output.. = bin/main/
 bin.includes = model/generated/,\
                model/,\
