+++ org.eclipse.xtext.xbase.testlanguages/src/org/eclipse/xtext/xbase/testlanguages/GenerateXbaseTestLanguages.mwe2
@@ -78,7 +78,9 @@ Workflow {
     			useDeprecatedClasses = false
     		}
     		fragment = types.TypesGeneratorFragment2 {}
-    		fragment = xbase.XbaseGeneratorFragment2 {}
+    		fragment = xbase.XbaseGeneratorFragment2 {
+    			generateXtendInferrer = false
+    		}
     		fragment = ConditionalXtextGeneratorFragment {
     			enabled = generateEclipse
     			fragment = ui.contentAssist.ContentAssistFragment2 {
@@ -115,7 +117,9 @@ Workflow {
     			useDeprecatedClasses = false
     		}
     		fragment = types.TypesGeneratorFragment2 {}
-    		fragment = xbase.XbaseGeneratorFragment2 {}
+    		fragment = xbase.XbaseGeneratorFragment2 {
+    			generateXtendInferrer = false
+    		}
     		
     		fragment = ConditionalXtextGeneratorFragment {
     			enabled = generateEclipse
@@ -154,7 +158,9 @@ Workflow {
     			useDeprecatedClasses = false
     		}
     		fragment = types.TypesGeneratorFragment2 {}
-    		fragment = xbase.XbaseGeneratorFragment2 {}
+    		fragment = xbase.XbaseGeneratorFragment2 {
+    			generateXtendInferrer = false
+    		}
     		
     		fragment = ConditionalXtextGeneratorFragment {
     			enabled = generateEclipse
