--- org.eclipse.xtext.xbase.testlanguages/src/org/eclipse/xtext/xbase/testlanguages/jvmmodel/XImportSectionTestLangJvmModelInferrer.xtend
@@ -1,13 +1,0 @@
-package org.eclipse.xtext.xbase.testlanguages.jvmmodel
-
-import org.eclipse.xtext.xbase.jvmmodel.AbstractModelInferrer
-
-/**
- * <p>Infers a JVM model from the source model.</p> 
- *
- * <p>The JVM model should contain all elements that would appear in the Java code 
- * which is generated from the source model. Other models link against the JVM model rather than the source model.</p>     
- */
-class XImportSectionTestLangJvmModelInferrer extends AbstractModelInferrer {
-}
-
