--- org.eclipse.xtext.xbase.testlanguages/xtend-gen/org/eclipse/xtext/xbase/testlanguages/bug462047/jvmmodel/Bug462047LangJvmModelInferrer.java
@@ -1,55 +1,0 @@
-package org.eclipse.xtext.xbase.testlanguages.bug462047.jvmmodel;
-
-import com.google.inject.Inject;
-import java.util.Arrays;
-import org.eclipse.emf.ecore.EObject;
-import org.eclipse.xtend2.lib.StringConcatenation;
-import org.eclipse.xtext.common.types.JvmGenericType;
-import org.eclipse.xtext.xbase.jvmmodel.AbstractModelInferrer;
-import org.eclipse.xtext.xbase.jvmmodel.IJvmDeclaredTypeAcceptor;
-import org.eclipse.xtext.xbase.jvmmodel.JvmTypesBuilder;
-import org.eclipse.xtext.xbase.lib.Extension;
-import org.eclipse.xtext.xbase.lib.Procedures.Procedure1;
-import org.eclipse.xtext.xbase.testlanguages.bug462047.bug462047.Bug462047Element;
-
-/**
- * <p>Infers a JVM model from the source model.</p>
- * 
- * <p>The JVM model should contain all elements that would appear in the Java code
- * which is generated from the source model. Other models link against the JVM model rather than the source model.</p>
- */
-@SuppressWarnings("all")
-public class Bug462047LangJvmModelInferrer extends AbstractModelInferrer {
-  @Inject
-  @Extension
-  private JvmTypesBuilder _jvmTypesBuilder;
-  
-  protected void _infer(final Bug462047Element element, final IJvmDeclaredTypeAcceptor acceptor, final boolean isPreIndexingPhase) {
-    String _name = element.getName();
-    String _plus = ("CORE." + _name);
-    final Procedure1<JvmGenericType> _function = (JvmGenericType it) -> {
-      StringConcatenation _builder = new StringConcatenation();
-      JvmGenericType _ref = element.getRef();
-      String _qualifiedName = null;
-      if (_ref!=null) {
-        _qualifiedName=_ref.getQualifiedName();
-      }
-      _builder.append(_qualifiedName);
-      this._jvmTypesBuilder.setDocumentation(it, _builder.toString());
-    };
-    acceptor.<JvmGenericType>accept(this._jvmTypesBuilder.toClass(element, _plus), _function);
-  }
-  
-  public void infer(final EObject element, final IJvmDeclaredTypeAcceptor acceptor, final boolean isPreIndexingPhase) {
-    if (element instanceof Bug462047Element) {
-      _infer((Bug462047Element)element, acceptor, isPreIndexingPhase);
-      return;
-    } else if (element != null) {
-      _infer(element, acceptor, isPreIndexingPhase);
-      return;
-    } else {
-      throw new IllegalArgumentException("Unhandled parameter types: " +
-        Arrays.<Object>asList(element, acceptor, isPreIndexingPhase).toString());
-    }
-  }
-}
